+++ src/test/converter2/renderer/index.ts
@@ -138,4 +138,5 @@ export interface DisabledGroups {
 export * as ExpandType from "./expandType";
 export * as GH2982 from "./gh2982";
 export { gh2995 } from "./gh2995";
+export * as GH3007 from "./gh3007";
 export { box as boxAlias };