+++ src/test/converter2/renderer/gh3014.ts
@@ -1,0 +1,7 @@
+import { GH3014Base } from "@typedoc/gh3014";
+
+export class GH3014 extends GH3014Base {
+    constructor() {
+        super();
+    }
+}