+++ src/test/converter2/renderer/index.ts
@@ -139,4 +139,5 @@ export * as ExpandType from "./expandType";
 export * as GH2982 from "./gh2982";
 export { gh2995 } from "./gh2995";
 export * as GH3007 from "./gh3007";
+export { GH3014 } from "./gh3014";
 export { box as boxAlias };